pyproject.toml
[project]
name = "fidgetsense-pc"
version = "0.1.0"
description = ""
authors = [
    {name = "ton1c",email = "sembl1@ya.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyserial (>=3.5,<4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
